# src/SpaTemplate.Infrastructure.Api/wwwroot/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v{version}/people-collections": {
      "post": {
        "tags": [
          "PeopleCollections"
        ],
        "summary": " ",
        "operationId": "PeopleCollections_CreateStudentCollection",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "studentForCreationDtos",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StudentForCreationDto"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CourseForCreationDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CourseForManipulationDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "CourseForManipulationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "StudentForCreationDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StudentForManipulationDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "courseForCreationDtos": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/CourseForCreationDto"
                }
              }
            }
          }
        ]
      },
      "StudentForManipulationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "isDone": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}